dpdk: join the raw socket threads before reading the final totals
txrx_wait_rx_workers() waits on the DPDK RX lcores only. The raw socket
ports run their TX and RX on pthreads, and those threads feed the same
totals - tx_targets[].stats is the raw leg's send side, dpdk_ext_rx_stats
is the external leg's return side. They were signalled to stop but not
joined until the teardown block sixty lines below the totals, so the
totals were read while they could still be holding unflushed counts.

For the raw RX workers this is a race, not a rare case: they exit on
force_quit_rx, which is set one statement before the totals are printed.
The comment above the call claimed "every counter is complete here",
which was only ever true of the DPDK workers.

stop_raw_socket_workers() is now idempotent and called before the
totals; the teardown block keeps its call and finds the work done. A
second pthread_join on the same thread is undefined behaviour, so
guarding the repeat rather than removing it keeps the error path that
reaches teardown without passing through the normal shutdown.

This does not explain the raw leg's 1,293 packet shortfall - both sides
of that comparison come from threads that had already finished (the raw
TX worker exits at Ctrl+C, ten seconds earlier, and the raw-origin
arrivals are counted by DPDK RX workers, which are waited for). It is a
separate defect found while checking whether the two subsystems are
counted consistently.

Also stops the sub-leg differences relying on unsigned wraparound to
print a negative.

+++ dpdk/src/Helpers.c
@@ -820,7 +820,7 @@ static void helper_render_final_totals(FILE *out,
                          (uint64_t)rte_atomic64_read(&dtn_stats[d].bad_pkts);
         }
         printf("    of that, DPDK TX workers  sent/back    : %lu / %lu  (%+ld)\n",
-               dpdk_sent, dpdk_back, (long)(dpdk_back - dpdk_sent));
+               dpdk_sent, dpdk_back, (long)((int64_t)dpdk_back - (int64_t)dpdk_sent));
         for (int r = 0; r < 2; r++) {
             struct raw_socket_port *rp = &raw_ports[r];
             uint64_t sent = 0, errs = 0;
@@ -833,7 +833,7 @@ static void helper_render_final_totals(FILE *out,
             printf("    of that, Port %-2u          sent/back    : %lu / %lu  (%+ld)"
                    ", send() errors %lu\n",
                    rp->port_id, sent, raw_back[r],
-                   (long)(raw_back[r] - sent), errs);
+                   (long)((int64_t)raw_back[r] - (int64_t)sent), errs);
         }
     }
     printf("  External TX leg (queue 4 -> Ports 12/13)\n");
